third_party/blink/web_tests/fast/forms/date-multiple-fields/date-multiple-fields-focus.html

<!DOCTYPE html>
<body>
<script src="../../../resources/js-test.js"></script>
<script>
description('Check if element.focus() does not focus on disabled sub-fields.');
if (window.internals)
    internals.runtimeFlags.langAttributeAwareFormControlUIEnabled = true;
else
    debug('This test requires DRT/WRT.');

var dateInput = document.createElement('input');
dateInput.lang = 'ja-jp';
dateInput.type = 'date';
dateInput.min = '2013-02-01';
dateInput.max = '2013-03-30';
document.body.appendChild(dateInput);
// Assume the date format is "yyyy/MM/dd"

dateInput.focus();
// Because the year field is fixed, the first focusable element is the
// month field.
shouldBeEqualToString('internals.shadowPseudoId(internals.shadowRoot(dateInput).activeElement)', '-webkit-datetime-edit-month-field');
dateInput.blur();
dateInput.focus();
shouldBeEqualToString('internals.shadowPseudoId(internals.shadowRoot(dateInput).activeElement)', '-webkit-datetime-edit-month-field');

dateInput.remove();
</script>
</body>
