third_party/blink/web_tests/platform/generic/fast/forms/date-multiple-fields/date-multiple-fields-focus-expected.txt

Check if element.focus() does not focus on disabled sub-fields.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.shadowPseudoId(internals.shadowRoot(dateInput).activeElement) is "-webkit-datetime-edit-month-field"
PASS internals.shadowPseudoId(internals.shadowRoot(dateInput).activeElement) is "-webkit-datetime-edit-month-field"
PASS successfullyParsed is true

TEST COMPLETE
